src/requirements.txt
selenium==4.7.2
webdriver-manager==3.8.5
python-telegram-bot==20.0a6
